[강운덕] [LUCYSUS-1744] spanId를 long으로 반납안하는 부분이 있어 수정.
git-svn-id: http://svn.bds.nhncorp.com/pe/hippo-tomcat-profiler/trunk@2746 84d0f5b1-2673-498c-a247-62c4ff18d310

+++ src/main/java/com/nhn/pinpoint/profiler/context/SpanId.java
@@ -14,7 +14,7 @@ public class SpanId {
     public static long newSpanId() {
         long id = seed.nextLong();
         while (id == NULL) {
-            id = seed.nextInt();
+            id = seed.nextLong();
         }
         return id;
     }
